follow-up r86195 (Sort of). Make sure the string there can really be exploded before calling explode on it.
I originally had the two variables that r86195 removed there to catch the case of a malformed
png file that didn't have a null byte where it was supposed to. However, just declaring the default
for those variables wouldn't have been enough to prevent the warning from list( blah, blah) = explode(..
So now explicitly check the null byte is there like it should be.

svn path=/trunk/phase3/; revision=86225