Remove the JPEG/TIFF specific metadata code from BitmapHandler and put it in JpegOrTiff handler to stop mostly irrelevent classes from getting it.
Also remove a method that is an exact duplicate of a base class (not sure whats with that).

This also coincidently fixes the issue with when a foreign file repo uses PagedTiffHandler
and the local one does not, and the builtin Tiff handler tries to treat the metadata as if
it was its own form.

svn path=/trunk/phase3/; revision=86567