(bug 28887) Do not re-use special page objects.
This causes bugs where if you do
{{special:recentchanges/3}} and {{special:recentchanges/15}} the first one is used for both
since the property is stored in a member variable of the class, which is then saved.

MediaWiki seems to have reused the special page classes since forever. I'm not sure why.
I can't imagine anyone relying on this behaviour, and it seems rather unexpected behaviour.

svn path=/trunk/phase3/; revision=87781