Followup to r86061: add parser test case to confirm that '&apos' in wikitext doesn't make it to output.
'&apos;' is not defined in HTML 4, and some browsers (for instance IE at least through 8) do not support it.
It is however part of XML 1 (and thus XHTML 1), and is also defined in HTML 5 provisional, so it's fine to accept it.

Normalization seems to already handle turning the &apos; named char reference into &#39; so we don't have any compatibility worries.

Also added a comment on the def in Sanitizer that it shouldn't be used in output.


svn path=/trunk/phase3/; revision=89681