Revert r91973 "	Bug 24254: Output JS var wgCategories on other actions than view as well"
This code just looks really wrong -- it seems to be stuffing things into OutputPage::getCategories() which seems clearly designed to only return a set of categories that have been previously set by something else; rather than a lazy loader it also seems to fetch them every time. Nor does it even check if the article exists before loading from the DB -- and in general it looks like it doesn't really fit in with other code. It also only lists itself as altering wgCategories, but it's changing what's in the OutputPage itself so this may affect other things.



svn path=/trunk/phase3/; revision=91974