Html.php: The "future"[1] is here. Add features for space-separated value attributes of html elements. * Has been suggested since August 2009 in r54767 (doc-comment from rawElement/element function) * Implements normalization for these attributes (removal of duplicates and redundant space) * Adds support for arrays (instead of just string) for these attributes. * String are still supported, and are converted to arrays to get the same normalization. * Wrote unit tests (which pass locally: $ php phpunit.php includes/HtmlTest.php) * Not trigger for the media-attribute. Reason: Although some people think it's space-separated, it's actually comma-separated. Treating them as space separated might even destroy the value. [2] [3]. Neither the html4 or html5 spec documents media-attribute as space-separated, and as of HTML5/CSS3 the media attribute may contain "media queries".
[1] "In the future, other HTML-specific features might be added, like allowing arrays for the values of attributes like class= and media=" in r54767 by Simetrical.
[2] http://www.w3.org/TR/1999/REC-html401-19991224/types.html#h-6.13
[3] http://dev.w3.org/csswg/css3-mediaqueries/#background

Implementation note: I choose to have a single list of attributes that trigger this feature. Some of these attributes only support multiple values and/or are documented as space-separated as of html5 (such as accesskey), but since those attributes in general have existed in html4 as well (just different w3c spec), they are not stripped if wgHtml5 is not true. So if this feature would (eg. for accesskey) would only be done if wgHtml5=true, then people could get output like <a accesskey=Array /> depending on a configuration variable, which will get messy and make developers' life hard.

svn path=/trunk/phase3/; revision=96170