bug 31990 justify paragraphs pref adds extra space to category listing
When text justification is enabled in user preference, the category list
could show a large space before the first category. Overriding the element
to use left justification is enough to fix the spacing issue albeit the
categories will no more justified.

Thanks to lupo with bug 31990 comments number 8 hinting at the catlinks
element not being properly justified.

This is a regression in REL1_18 and need a backport.



svn path=/trunk/phase3/; revision=102297