The URL wasn't loading when the database is used (interwiki cache works). "PHP Notice:  Trying to get property of non-object in /www/w/includes/api/ApiQuerySiteinfo.php on line 295"
This reverts r97806. It took me some time to find out, but r96475 accidentally backported the change from $row->iw_url to $row['iw_url'] (etc) to 1.18, which was fixed in r97789. That revision was merged to trunk, which shouldn't have been done either. I.e. until 1.18 it is an object, from 1.19 it is an array.

svn path=/trunk/phase3/; revision=104573